feat: move login form header to folder and add test
